web/xanadus-scripted-symphony/setup/users-service/requirements.txt
Flask>=2.2.0
SQLAlchemy>=1.4.0
Flask-SQLAlchemy==3.0.3
Werkzeug>=3.0.0
flask-login>=0.6.3
pillow>=10.3.0
requests>=2.32.0
